Disable course number, term, and instructor dropdowns
When full functionality of the dropdowns has been implemented, the course number dropdown are disabled until the user has chosen a subject to search. The term dropdown will be enabled when the user choose a course number, and the instructor dropdown will be enabled when the user chooses a term.